Create config.mini for fbterm/linux console cols
This is for parsing with https://gist.github.com/ShadowKyogre/e4672b7708191fd8ae71

FBTerm and the Linux CLI can't specify special default background colors and 
foreground colors separate from the palette. What this does is that it changes the
blue-green "blacks" to the purple background and replaces the blueish whites with
warmer whites. Will need to test later when I get back on my Linux half.